requirements.txt
pytorch \
torchvision \
cudatoolkit=10.0 \
h5py \
tqdm \
-c pytorch